pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Matthew Baxter <matthewbaxter0412@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-postgres = "^0.0.9"
torch = {version = "^2.4.0", python = "^3.11", platform = "macos"}
langchain-community = "^0.2.15"
gradio = "^4.42.0"
python-dotenv = "^1.0.1"
black = "^24.8.0"
pypdf = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
